+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # ChangeLog
 
+* Trait recipe handling io can be BOTH now.
+* Bump up ldlib.
+* Add Search Component for proxy recipe selection.
 * Add the geckolib glowing layer for emissive rendering.
 * Add proxy supports for mbd2 recipetypes.
 * Add XEI supports for fuel recipes.
 * Add more events related to the fuel handling.
   * MachineFuelBurningFinishEvent
   * MachineFuelRecipeModifyEvent
-* Trait recipe handling io can be BOTH now.
-* Bump up ldlib